Skeleton Upload
This is the upload of the skeleton of my website. It includes the rough html layout, the navigation bar, landing, about information and placeholder slides. I've also included styles The organize the pages universally and more specific styles for each section. Lastly, I've implemented scripts to control the website color, the taglines, and the projects slideshow.